refill script
